oracle/analysis/requirements.txt
fastapi
uvicorn
pandas
numpy
scipy
statsmodels
python-docx
python-multipart
jinja2
scikit-posthocs
networkx
matplotlib
seaborn
openpyxl
scikit-learn
plotly
slowapi
python-dotenv
